update: putnbr_fd

+++ ft_putnbr_fd.c
@@ -6,7 +6,7 @@
 /*   By: jmanani <jmanani@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/18 15:58:53 by jmanani           #+#    #+#             */
-/*   Updated: 2025/10/23 20:43:53 by jmanani          ###   ########.fr       */
+/*   Updated: 2025/10/23 20:46:24 by jmanani          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ void	ft_putnbr_fd(int n, int fd)
 		num = -num;
 	}
 	if (num >= 10)
-		ft_putnbr_fd(num / 10, fd);
+		ft_putnbr_fd((int)(num / 10), fd);
 	print_char = (num % 10) + 48;
 	write(fd, &print_char, 1);
 }
